the first 50 Fibonacci numbers, starting with 1 and 2, followed by a new line

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -1,17 +1,17 @@
 #include "main.h"
 /**
  ** main - Entry point
- * 
+ *
  * Return: Always 0 (Success)
  */
 int main(void)
 {
-	int i, f1 = 0, f2 = 1, s;
+	unsigned long i, f1 = 0, f2 = 1, s;
 
 	for (i = 0; i < 50; i++)
 	{
 		s = f1 + f2;
-		printf("%d", s);
+		printf("%lu", s);
 		f1 = f2;
 		f2 = s;
 		if (i == 49)
